🚚 rename: Comment 분리
