+++ src/lib/comment/hooks/useDeleteComment.tsx
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { deleteComment } from "../../api";
-import { COMMENT, POST } from "../../key";
+import { deleteComment } from '../api';
+import { COMMENT } from '../key';
+import { POST } from '@/lib/post/key';
 
 export default function useDeleteComment(postId: string, commentId: string) {
   const queryClient = useQueryClient();
