+++ src/lib/comment/key.ts
@@ -1,0 +1,1 @@
+export const COMMENT = "COMMENT";
