+++ src/lib/post/key.ts
@@ -3,5 +3,3 @@ export const POST = "POST";
 export const POST_LIKE_STATUS = "POST_LIKE_STATUS";
 export const POST_LIKE_COUNT = "POST_LIKE_COUNT";
 export const POST_LIKED_BY_USER = "POST_LIKED_BY_USER";
-
-export const COMMENT = "COMMENT";
